crypto: hisilicon/sec - modify the hardware endian
From a52626106d6f7edf3d106c065e13a0313cfeb82f Mon Sep 17 00:00:00 2001
 configuration
Git-commit: a52626106d6f7edf3d106c065e13a0313cfeb82f
Patch-mainline: v5.15-rc1
References: bsn#12

When the endian configuration of the hardware is abnormal, it will
cause the SEC engine is faulty that reports empty message. And it
will affect the normal function of the hardware. Currently the soft
configuration method can't restore the faulty device. The endian
needs to be configured according to the system properties. So fix it.

Signed-off-by: Kai Ye <yekai13@huawei.com>
Signed-off-by: Herbert Xu <herbert@gondor.apana.org.au>
Signed-off-by: Guoqing Jiang <guoqing.jiang@suse.com>

Patch-name: patches.mainline/v5.15-rc1-crypto-hisilicon-sec-modify-the-hardware-endian-conf.patch