mptcp: more accurate timeout
Git-commit: 33d41c9cd74c56643a710810703b6bce6eb25efa
Message-Id: <33d41c9cd74c56643a710810703b6bce6eb25efa.1637072199.git.geliang.tang@suse.com>
References: bsn#2
Patch-mainline: v5.15-rc1

As reported by Maxim, we have a lot of MPTCP-level
retransmissions when multilple links with different latencies
are in use.

This patch refactor the mptcp-level timeout accounting so that
the maximum of all the active subflow timeout is used. To avoid
traversing the subflow list multiple times, the update is
performed inside the packet scheduler.

Additionally clean-up a bit timeout handling.

Signed-off-by: Paolo Abeni <pabeni@redhat.com>
Signed-off-by: Mat Martineau <mathew.j.martineau@linux.intel.com>
Signed-off-by: David S. Miller <davem@davemloft.net>
Signed-off-by: Geliang Tang <geliang.tang@suse.com>

Patch-name: patches.suse/v5.15-rc1-mptcp-more-accurate-timeout.patch