scsi: lpfc: Add support for cm enablement buffer
From 72df8a452883b0be334396acba07df77c3c3f6c7 Mon Sep 17 00:00:00 2001
Git-commit: 72df8a452883b0be334396acba07df77c3c3f6c7
Patch-mainline: v5.15-rc1
References: bsn#61

As part of the cmf framework, the firmware maintains a table with
congestion related state information, specifically whether enabled and if
enabled, whether monitoring or actively managing congestion.

Add definition of the table and add support to read the table from the
adapter and determine if it is enabled. In support of this, the READ_OBJECT
mailbox command definition is added to the driver.

Link: https://lore.kernel.org/r/20210816162901.121235-8-jsmart2021@gmail.com
Co-developed-by: Justin Tee <justin.tee@broadcom.com>
Signed-off-by: Justin Tee <justin.tee@broadcom.com>
Signed-off-by: James Smart <jsmart2021@gmail.com>
Signed-off-by: Martin K. Petersen <martin.petersen@oracle.com>
Signed-off-by: Guoqing Jiang <guoqing.jiang@suse.com>

Patch-name: patches.suse/v5.15-rc1-scsi-lpfc-Add-support-for-cm-enablement-buffer.patch