f2fs: guarantee to write dirty data when enabling checkpoint
Git-commit: 611b7f9dc9f696ba0f9ab03b5c6500f3718cbebe
 back
References: bsn#19
Patch-mainline: v5.10.65

commit dddd3d65293a52c2c3850c19b1e5115712e534d8 upstream.

We must flush all the dirty data when enabling checkpoint back. Let's guarantee
that first by adding a retry logic on sync_inodes_sb(). In addition to that,
this patch adds to flush data in fsync when checkpoint is disabled, which can
mitigate the sync_inodes_sb() failures in advance.

Reviewed-by: Chao Yu <chao@kernel.org>
Signed-off-by: Jaegeuk Kim <jaegeuk@kernel.org>
Signed-off-by: Greg Kroah-Hartman <gregkh@linuxfoundation.org>
Signed-off-by: Kai Liu <kai.liu@suse.com>

Patch-name: patches.stable/v5.10.65-0217-f2fs-guarantee-to-write-dirty-data-when-enabling-che.patch