mptcp: send out MP_FAIL when data checksum fails
Git-commit: 478d770008b03ed9d74bdc8add2315b7fd124ecc
Message-Id: <478d770008b03ed9d74bdc8add2315b7fd124ecc.1637071919.git.geliang.tang@suse.com>
References: bsn#2
Patch-mainline: v5.15-rc1

When a bad checksum is detected, set the send_mp_fail flag to send out
the MP_FAIL option.

Add a new function mptcp_has_another_subflow() to check whether there's
only a single subflow.

When multiple subflows are in use, close the affected subflow with a RST
that includes an MP_FAIL option and discard the data with the bad
checksum.

Set the sk_state of the subsocket to TCP_CLOSE, then the flag
MPTCP_WORK_CLOSE_SUBFLOW will be set in subflow_sched_work_if_closed,
and the subflow will be closed.

When a single subfow is in use, temporarily handled by sending MP_FAIL
with a RST too.

Signed-off-by: Geliang Tang <geliangtang@xiaomi.com>
Signed-off-by: Mat Martineau <mathew.j.martineau@linux.intel.com>
Signed-off-by: David S. Miller <davem@davemloft.net>
Signed-off-by: Geliang Tang <geliang.tang@suse.com>

Patch-name: patches.suse/v5.15-rc1-mptcp-send-out-MP_FAIL-when-data-checksum-fails.patch