RDMA/hns: Bugfix for incorrect association between dip_idx
From eb653eda1e91dd3e7d1d2448d528d033dbfbe78f Mon Sep 17 00:00:00 2001
 and dgid
Git-commit: eb653eda1e91dd3e7d1d2448d528d033dbfbe78f
Patch-mainline: v5.15-rc1
References: bsn#12

dip_idx and dgid should be a one-to-one mapping relationship, but when
qp_num loops back to the start number, it may happen that two different
dgid are assiociated to the same dip_idx incorrectly.

One solution is to store the qp_num that is not assigned to dip_idx in an
array. When a dip_idx needs to be allocated to a new dgid, an spare qp_num
is extracted and assigned to dip_idx.

Fixes: f91696f2f053 ("RDMA/hns: Support congestion control type selection according to the FW")
Link: https://lore.kernel.org/r/1629884592-23424-4-git-send-email-liangwenpeng@huawei.com
Signed-off-by: Junxian Huang <huangjunxian4@hisilicon.com>
Signed-off-by: Yangyang Li <liyangyang20@huawei.com>
Signed-off-by: Wenpeng Liang <liangwenpeng@huawei.com>
Signed-off-by: Jason Gunthorpe <jgg@nvidia.com>
Signed-off-by: Guoqing Jiang <guoqing.jiang@suse.com>

Patch-name: patches.mainline/v5.15-rc1-RDMA-hns-Bugfix-for-incorrect-association-between-di.patch